Add initial fix for tool being broken
Tool failed because the stopwords couldn't be loaded.

This is because it uses nltk to retrieve the data every time.

The nltk_data had been removed from the chroot in
https://github.com/scraperwiki/salt-data/commit/756563d3a1e4a1af5e473541a53440d2511c86dd

Further to that, even retrieving it directly using nltk in boxes
seemed to fail with a urllib2 error (unless I used a virtualenv
and installed nltk there).

So, the simplest fix seems to be to use a static list of words.

NB: nltk is distributed under an Apache 2.0 licence but:

"NLTK corpora are provided under the terms given in the README file
for each corpus; all are redistributable, and available for
non-commercial use."